Replace dependency execSync module with sync-exec
ds18x20 cannot be currently installed with node.js 0.12 as
execSync module doesn't work with nodejs 0.12 and the author of
that module cannot make an update to the module to fix that. [1]

We replace execSync with sync-exec [2] which is "an fs.execSync
replacement until you get it natively from node 0.12+".

Also, the error number for ENOENT in nodejs 0.12 is -2 instead of
34 which is was previously. Therefore we remove the check
"errno == 34" (it should be anyways enough to test that
"code == ENOENT").

[1] https://github.com/mgutz/execSync/issues/38
[2] https://www.npmjs.com/package/sync-exec